refactor(topology_cleanup): extract generic graph-topology reroot operations from commands/clock/
Move EdgeSplitInfo, EdgeMergeInfo, RerootResult, RerootChanges data types and generic graph operations (split_edge, apply_reroot_topology, remove_node_if_trivial) from commands/clock/reroot.rs to representation/algo/topology_cleanup/reroot.rs.

- Generic operations now require only GraphNode + GraphEdge + HasBranchLength bounds (no clock dependency)
- clock/reroot.rs becomes a thin wrapper: reroot_in_place delegates to shared topology functions, then applies clock-specific edge message updates
- All consumers (timetree, partition, tests) updated to import topology types from topology_cleanup
- Re-exports from clock/reroot.rs preserve backward compatibility for clock-internal consumers
