+++ packages/treetime/src/commands/clock/__tests__/mod.rs
@@ -1,6 +1,6 @@
 mod test_clock_dengue100;
-mod test_clock_set;
 mod test_clock_filter;
 mod test_clock_regression;
+mod test_clock_set;
 mod test_date_constraints;
 mod test_reroot;
