+++ packages/treetime/src/commands/clock/reroot.rs
@@ -4,13 +4,18 @@ use crate::commands::clock::clock_traits::{ClockEdge, ClockNode};
 use crate::commands::clock::find_best_root::find_best_root::find_best_root;
 use crate::commands::clock::find_best_root::find_best_split::FindRootResult;
 use crate::commands::clock::find_best_root::params::BranchPointOptimizationParams;
+use crate::representation::algo::topology_cleanup::reroot::{
+  self as topology_reroot, remove_node_if_trivial, split_edge,
+};
 use approx::ulps_eq;
 use eyre::Report;
 use serde::{Deserialize, Serialize};
 use smart_default::SmartDefault;
-use treetime_graph::edge::{GraphEdge, GraphEdgeKey, invert_edge};
+use treetime_graph::edge::{GraphEdge, GraphEdgeKey};
 use treetime_graph::graph::Graph;
 use treetime_graph::node::{GraphNode, GraphNodeKey};
+
+pub use topology_reroot::{EdgeMergeInfo, EdgeSplitInfo, RerootChanges, RerootResult};
 
 /// Controls reroot behavior for graph topology changes.
 #[derive(Clone, Debug, Serialize, Deserialize, SmartDefault)]
@@ -30,59 +35,6 @@ pub struct RerootParams {
   /// Use false for pre-filter steps where outliers may cause negative rates at all positions.
   #[default = true]
   pub force_positive_rate: bool,
-}
-
-/// Information about an edge split during reroot.
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct EdgeSplitInfo {
-  /// The original edge that was split.
-  pub old_edge_key: GraphEdgeKey,
-  /// The new node created at the split point.
-  pub new_node_key: GraphNodeKey,
-  /// The edge from the original source to the new node.
-  pub parent_side_edge_key: GraphEdgeKey,
-  /// The edge from the new node to the original target.
-  pub child_side_edge_key: GraphEdgeKey,
-  /// Position along the edge where the split occurred (0.0 = target, 1.0 = source).
-  pub split_position: f64,
-}
-
-/// Information about an edge merge when removing a trivial node.
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct EdgeMergeInfo {
-  /// The node that was removed.
-  pub removed_node_key: GraphNodeKey,
-  /// The edge from parent to the removed node.
-  pub parent_edge_key: GraphEdgeKey,
-  /// The edge from the removed node to its child.
-  pub child_edge_key: GraphEdgeKey,
-  /// The new merged edge from parent to child.
-  pub merged_edge_key: GraphEdgeKey,
-}
-
-/// Result of a reroot operation.
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct RerootResult {
-  /// The key of the new root node.
-  pub new_root_key: GraphNodeKey,
-  /// Information about edge split, if one occurred.
-  pub edge_split: Option<EdgeSplitInfo>,
-  /// Information about edge merge, if a trivial node was removed.
-  pub edge_merge: Option<EdgeMergeInfo>,
-}
-
-/// Bundles all topology changes from a reroot operation for partition updates.
-///
-/// Passed to `PartitionRerootOps::apply_reroot` to update partition state in a single call.
-#[derive(Clone, Debug, Default)]
-pub struct RerootChanges {
-  /// Edge split info if a new node was created at the reroot point.
-  pub edge_split: Option<EdgeSplitInfo>,
-  /// Edge merge info if the old root was removed as a trivial node.
-  pub edge_merge: Option<EdgeMergeInfo>,
-  /// Keys of edges on the path from old root to new root (post-inversion direction).
-  /// Empty if root did not change or old root was removed.
-  pub inverted_edge_keys: Vec<GraphEdgeKey>,
 }
 
 pub fn reroot_in_place<N, E, D>(
@@ -110,7 +62,7 @@ where
     });
   };
 
-  // Extract edge endpoints before potentially removing the edge
+  // Extract edge endpoints before the edge is removed by split_edge
   let (source_key, target_key) = {
     let edge = graph.get_edge(edge_key).expect("Edge not found");
     (edge.read_arc().source(), edge.read_arc().target())
@@ -136,8 +88,6 @@ where
     apply_reroot(graph, old_root_key, new_root_key, options)?;
 
     if reroot_params.remove_trivial_root {
-      // Clean up old root if it is now a trivial node, i.e. has exactly one parent and one child
-      // Nb: the attributes of the old root node and branches (other than branch lengths) are discarded
       remove_node_if_trivial(graph, old_root_key)?
     } else {
       None
@@ -153,8 +103,7 @@ where
   })
 }
 
-/// Create new root node by splitting the edge into two.
-/// Returns the split info containing the new node and edge keys.
+/// Create new root node by splitting the edge into two, then setting clock data on the new node.
 fn create_new_root_node<N, E, D>(
   graph: &mut Graph<N, E, D>,
   edge_key: GraphEdgeKey,
@@ -166,39 +115,19 @@ where
   E: GraphEdge + ClockEdge + Default,
   D: Send + Sync,
 {
-  let mut new_node = N::default();
-  new_node.set_clock_set(clock_set);
-  let new_node_key = graph.add_node(new_node);
+  let split_info = split_edge(graph, edge_key, split)?;
 
-  // Extract edge data before removing the edge to avoid Arc reference issues
-  let (source_key, target_key, branch_length) = {
-    let edge = graph.get_edge(edge_key).expect("Edge not found");
-    let source_key = edge.read_arc().source();
-    let target_key = edge.read_arc().target();
-    let branch_length = edge.read_arc().payload().read_arc().branch_length().unwrap_or_default();
-    (source_key, target_key, branch_length)
-  };
+  // Set clock-specific data on the new split node
+  let new_node = graph
+    .get_node(split_info.new_node_key)
+    .expect("New split node not found");
+  new_node.write_arc().payload().write_arc().set_clock_set(clock_set);
 
-  let mut parent_edge = E::default();
-  parent_edge.set_branch_length(Some(split * branch_length));
-  let parent_side_edge_key = graph.add_edge(source_key, new_node_key, parent_edge)?;
-
-  let mut child_edge = E::default();
-  child_edge.set_branch_length(Some((1.0 - split) * branch_length));
-  let child_side_edge_key = graph.add_edge(new_node_key, target_key, child_edge)?;
-
-  graph.remove_edge(edge_key)?;
-
-  Ok(EdgeSplitInfo {
-    old_edge_key: edge_key,
-    new_node_key,
-    parent_side_edge_key,
-    child_side_edge_key,
-    split_position: split,
-  })
+  Ok(split_info)
 }
 
-/// Modify graph topology to make the newly identified root the actual root.
+/// Modify graph topology to make the newly identified root the actual root,
+/// then update clock-specific edge messages.
 fn apply_reroot<N, E, D>(
   graph: &mut Graph<N, E, D>,
   old_root_key: GraphNodeKey,
@@ -210,79 +139,20 @@ where
   E: GraphEdge + ClockEdge,
   D: Send + Sync,
 {
-  // Find paths from the old root to the new desired root
-  let paths = graph.path_from_node_to_node(new_root_key, old_root_key)?;
+  // Invert edges along the path (generic topology operation)
+  let inverted_edge_keys = topology_reroot::apply_reroot_topology(graph, old_root_key, new_root_key)?;
 
-  // Invert every edge on the path from old to new root.
-  // This will make the desired new root into an actual root. The old root might no longer be a root.
-  for (_, edge) in &paths {
-    if let Some(edge) = edge {
-      invert_edge(graph, edge);
-      let mut edge_payload = edge.read_arc().payload().write_arc();
-      let edge_len = edge_payload.branch_length().unwrap();
-      let branch_variance = options.variance_factor * edge_len + options.variance_offset;
-      let tmp_to_parent = edge_payload.to_parent().clone();
-      *edge_payload.to_parent_mut() = edge_payload.to_child().clone();
-      *edge_payload.to_child_mut() = tmp_to_parent;
-      *edge_payload.from_child_mut() = edge_payload.to_parent().propagate_averages(edge_len, branch_variance);
-    }
+  // Update clock-specific edge messages for inverted edges
+  for edge_key in &inverted_edge_keys {
+    let edge = graph.get_edge(*edge_key).expect("Inverted edge not found");
+    let mut edge_payload = edge.read_arc().payload().write_arc();
+    let edge_len = edge_payload.branch_length().unwrap();
+    let branch_variance = options.variance_factor * edge_len + options.variance_offset;
+    let tmp_to_parent = edge_payload.to_parent().clone();
+    *edge_payload.to_parent_mut() = edge_payload.to_child().clone();
+    *edge_payload.to_child_mut() = tmp_to_parent;
+    *edge_payload.from_child_mut() = edge_payload.to_parent().propagate_averages(edge_len, branch_variance);
   }
 
-  // Some bookkeeping
-  graph.build()?;
   Ok(())
-}
-
-/// Remove a node if it is trivial (i.e. has exactly one parent and one child), merging the edges.
-/// Returns `Some(EdgeMergeInfo)` if the node was removed, `None` if the node was not trivial.
-/// Nb: when applied this function removes properties of the removed node and its connecting edges.
-pub(super) fn remove_node_if_trivial<N, E, D>(
-  graph: &mut Graph<N, E, D>,
-  node_key: GraphNodeKey,
-) -> Result<Option<EdgeMergeInfo>, Report>
-where
-  N: GraphNode + ClockNode,
-  E: GraphEdge + ClockEdge + Default,
-  D: Send + Sync,
-{
-  let (parent_edge_key, child_edge_key) = {
-    let node = graph.get_node(node_key).expect("Node not found");
-    let node = node.read_arc();
-    if node.inbound().len() != 1 || node.outbound().len() != 1 {
-      return Ok(None);
-    }
-    (node.inbound()[0], node.outbound()[0])
-  };
-
-  let (parent_key, parent_branch) = {
-    let parent_edge = graph.get_edge(parent_edge_key).expect("Parent edge not found");
-    let parent_edge = parent_edge.read_arc();
-    (parent_edge.source(), parent_edge.payload().read_arc().branch_length())
-  };
-
-  let (child_key, child_branch) = {
-    let child_edge = graph.get_edge(child_edge_key).expect("Child edge not found");
-    let child_edge = child_edge.read_arc();
-    (child_edge.target(), child_edge.payload().read_arc().branch_length())
-  };
-
-  let merged_branch_length = match (parent_branch, child_branch) {
-    (Some(a), Some(b)) => Some(a + b),
-    (a, b) => a.or(b),
-  };
-
-  graph.remove_node(node_key)?;
-
-  let mut merged_payload = E::default();
-  merged_payload.set_branch_length(merged_branch_length);
-  let merged_edge_key = graph.add_edge(parent_key, child_key, merged_payload)?;
-
-  graph.build()?;
-
-  Ok(Some(EdgeMergeInfo {
-    removed_node_key: node_key,
-    parent_edge_key,
-    child_edge_key,
-    merged_edge_key,
-  }))
 }
