+++ packages/treetime/src/commands/timetree/optimization/reroot.rs
@@ -2,8 +2,9 @@ use crate::commands::ancestral::marginal::update_marginal;
 use crate::commands::clock::clock_model::ClockModel;
 use crate::commands::clock::clock_regression::{ClockParams, estimate_clock_model_with_reroot_policy};
 use crate::commands::clock::find_best_root::params::BranchPointOptimizationParams;
-use crate::commands::clock::reroot::{RerootChanges, RerootParams};
+use crate::commands::clock::reroot::RerootParams;
 use crate::commands::timetree::partition_ops::PartitionTimetreeAll;
+use crate::representation::algo::topology_cleanup::reroot::RerootChanges;
 use crate::representation::partition::timetree::GraphTimetree;
 use crate::representation::payload::timetree::EdgeTimetree;
 use crate::representation::payload::timetree::NodeTimetree;
