+++ packages/treetime/src/representation/algo/topology_cleanup/mod.rs
@@ -2,3 +2,4 @@
 mod __tests__;
 
 pub mod collapse;
+pub mod reroot;
